requirements.txt
mcp>=0.9.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
